lntest: Move simnet miner to testutils pkg
This moves the AdjustedSimnetMiner function from the lntest package to
the internal/testutils package. This lets this function be used in
packages which would otherwise cause a circular package dependency
error.
