+++ internal/testutils/simnet_miner.go
@@ -1,4 +1,4 @@
-package lntest
+package testutils
 
 import (
 	"context"
